\name{Deprecated functions}
\alias{add_actors_ml}
\alias{add_actors_ml_v3}
\alias{add.attributes.ml}
\alias{attributes.ml}
\alias{get.values.ml}
\alias{set.values.ml}
\alias{new.attributes.ml}
\alias{list.attributes.ml}
\alias{abacus.ml}
\alias{clique.percolation.ml}
\alias{glouvain.ml}
\alias{infomap.ml}
\alias{get.community.list.ml}
\alias{set.directed.ml}
\alias{is.directed.ml}
\alias{distance.ml}
\alias{grow.ml}
\alias{evolution.pa.ml}
\alias{evolution.er.ml}
\alias{write.ml}
\alias{layer.summary.ml}
\alias{layer.comparison.ml}
\alias{layout.multiforce.ml}
\alias{layout.circular.ml}
\alias{add.layers.ml}
\alias{add.actors.ml}
\alias{add.vertices.ml}
\alias{add.edges.ml}
\alias{add.igraph.layer.ml}
\alias{delete.layers.ml}
\alias{delete.actors.ml}
\alias{delete.vertices.ml}
\alias{delete.edges.ml}
\alias{neighbors.ml}
\alias{xneighbors.ml}
\alias{degree.ml}
\alias{degree.deviation.ml}
\alias{neighborhood.ml}
\alias{xneighborhood.ml}
\alias{connective.redundancy.ml}
\alias{relevance.ml}
\alias{xrelevance.ml}
\alias{ml.empty}
\alias{ml.aucs}
\alias{ml.bankwiring}
\alias{ml.florentine}
\alias{ml.monastery}
\alias{ml.tailorshop}
\alias{ml.toy}
\alias{layers.ml}
\alias{actors.ml}
\alias{vertices.ml}
\alias{edges.ml}
\alias{edges.idx.ml}
\alias{num.layers.ml}
\alias{num.actors.ml}
\alias{num.vertices.ml}
\alias{num.edges.ml}
\alias{flatten.ml}
\alias{add.nodes.ml}
\alias{delete.nodes.ml}
\alias{nodes.ml}
\alias{num.nodes.ml}
\alias{read.ml}
\title{
Deprecated functions
}
\description{
These functions are deprecated and will be removed at the next major version upgrade.
}
\usage{
add_actors_ml(...)
add_actors_ml_v3(n, actors)
add.attributes.ml(...)
attributes.ml(...)
get.values.ml(...)
set.values.ml(...)
new.attributes.ml(...)
list.attributes.ml(...)
abacus.ml(...)
clique.percolation.ml(...)
glouvain.ml(...)
infomap.ml(...)
get.community.list.ml(...)
set.directed.ml(...)
is.directed.ml(...)
distance.ml(...)
grow.ml(...)
evolution.pa.ml(...)
evolution.er.ml(...)
read.ml(...)
write.ml(...)
read.ml(...)
layer.summary.ml(...)
layer.comparison.ml(...)
layout.multiforce.ml(...)
layout.circular.ml(...)
add.layers.ml(...)
add.actors.ml(...)
add.vertices.ml(...)
add.edges.ml(...)
add.igraph.layer.ml(...)
delete.layers.ml(...)
delete.actors.ml(...)
delete.vertices.ml(...)
delete.edges.ml(...)
neighbors.ml(...)
xneighbors.ml(...)
degree.ml(...)
degree.deviation.ml(...)
neighborhood.ml(...)
xneighborhood.ml(...)
connective.redundancy.ml(...)
relevance.ml(...)
xrelevance.ml(...)
ml.empty(...)
ml.aucs(...)
ml.bankwiring(...)
ml.florentine(...)
ml.monastery(...)
ml.tailorshop(...)
ml.toy(...)
layers.ml(...)
actors.ml(...)
vertices.ml(...)
edges.ml(...)
edges.idx.ml(...)
num.layers.ml(...)
num.actors.ml(...)
num.vertices.ml(...)
num.edges.ml(...)
flatten.ml(...)
add.nodes.ml(...)
delete.nodes.ml(...)
nodes.ml(...)
num.nodes.ml(...)
}
\arguments{
\item{...}{deprecated}
\item{n}{A multilayer network.}
\item{actors}{An array of names of actors.}
}
