% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Generating the simulated data.}
\usage{
generate.data(N, Mu0.list, Theta0.list, Sigma0.list)
}
\arguments{
\item{N}{@ N: K0 * 1 vector, the sample sizes of subgroups.}

\item{Mu0.list}{a list including K0 mean vectors (p * 1).}

\item{Theta0.list}{a list including K0 precision matrices (p * p).}

\item{Sigma0.list}{a list including K0 correlation matrices (p * p).}
}
\value{
A list "whole.data" including:
\item{L0}{n * 1 vector, the subgroup labels to which each sample belongs.}
\item{Mu0}{K0 * p matrix, K0 mean vectors.}
\item{Theta0}{K0 * p * p array, K0 precision matrices.}
\item{data}{n * p matrix, the design matrix.}
\item{n_all}{int, the total sample size.}
\item{K0}{int, the true number of subgroups.}
}
\description{
Generating the simulated data.
}
