% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.law.network.R
\name{Power.law.network}
\alias{Power.law.network}
\title{Generating the s-block power-law precision matrices.}
\usage{
Power.law.network(p, s = 10, umin = 0.4, umax = 0.7, num.differ)
}
\arguments{
\item{p}{Dimensions of the precision matrix.}

\item{s}{The number of sub-networks.}

\item{umin}{The lower bound of non-zero elements on non-diagonal elements.}

\item{umax}{The upper bound of non-zero elements on non-diagonal elements.}

\item{num.differ}{number of sub-networks which are different across subgroups.}
}
\value{
A list including The precision matrices of three subgroups.
}
\description{
Generating the s-block power-law precision matrices.
}
