% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitGMM.R
\name{refitGMM}
\alias{refitGMM}
\title{Refit function}
\usage{
refitGMM(X, Theta, mu, L.mat, A.hat)
}
\arguments{
\item{X}{An n by p data matrix.}

\item{Theta}{A p-by-p-by-K array of K precision matrices.}

\item{mu}{A K by p matrix of K cluster mean.}

\item{L.mat}{A n by K matrix. Element (i, j) represents whether the i-th sample belongs to the j-th category.}

\item{A.hat}{A p-by-p-by-K array of K edge matrices.}
}
\value{
A list containing the precision matrices array and mean.
}
\description{
Refit function
}
